Abstracting tracker comms
