+++ crates/librqbit/src/lib.rs
@@ -36,7 +36,8 @@ mod session;
 mod spawn_utils;
 mod torrent_state;
 pub mod tracing_subscriber_config_utils;
-mod tracker_comms;
+pub mod tracker_comms;
+pub mod tracker_comms_http;
 pub mod tracker_comms_udp;
 mod type_aliases;
